+++ seqism/seqism-common/src/main/java/dev/seqism/common/vo/SeqismMessage.java
@@ -41,39 +41,4 @@ public class SeqismMessage<T> implements Serializable {
     static <T> SeqismMessage<T> of(SeqismMessageHeader header, T body) {
         return new SeqismMessage<T>(header, body);
     }
-
-    @Data
-    @AllArgsConstructor
-    public static class SeqismMessageHeader implements Serializable {
-        private final String bizCode;
-        private final String tranId;
-        private final SeqismMessageStatus status;
-        private final ErrorInfo error;
-
-        public SeqismMessageHeader() {
-            this(null, null, SeqismMessageStatus.FAILURE, null);
-        }
-
-        SeqismMessageHeader toInProgress() {
-            return toInProgress(tranId);
-        }
-
-        SeqismMessageHeader toInProgress(String tranId) {
-            return new SeqismMessageHeader(bizCode, tranId, SeqismMessageStatus.IN_PROGRESS, null);
-        }
-
-        SeqismMessageHeader toSuccess() {
-            return new SeqismMessageHeader(bizCode, tranId, SeqismMessageStatus.SUCCESS, null);
-        }
-
-        SeqismMessageHeader toFailure(ErrorInfo errorInfo) {
-            return new SeqismMessageHeader(bizCode, tranId, SeqismMessageStatus.FAILURE, errorInfo);
-        }
-    }
-
-    public enum SeqismMessageStatus {
-        IN_PROGRESS,   // 진행중
-        SUCCESS,       // 완료_성공
-        FAILURE        // 완료_실패
-    }
 }
