+++ seqism/seqism-common/src/main/java/dev/seqism/common/vo/SeqismMessageStatus.java
@@ -1,0 +1,7 @@
+package dev.seqism.common.vo;
+
+public enum SeqismMessageStatus {
+    IN_PROGRESS,   // 진행중
+    SUCCESS,       // 완료_성공
+    FAILURE        // 완료_실패
+}
